refactor: split architecture-map interactions into focused modules
ArchitectureMap.tsx was 972 lines, dominated by a single 454-line useEffect
that wired pan, zoom, drag, keyboard, focus-trap and a11y announcements by
reaching into prerendered DOM through selectors.

The effect decomposes into lib/architecture-map: geometry.ts holds the bezier
and viewport math as pure functions taking rects instead of reading the DOM,
filter.ts takes the filter as a parameter instead of closing over a mutable,
viewport.ts owns the transform state with wheel and drag, keyboard.ts owns the
shortcut switch and the help-dialog focus trap, and interactions.ts composes
them behind attachMap(root, { onPanel }). Zoom factors, scale clamps and arrow
steps are now named constants.

Markup splits into Sidebar.tsx, Layers.tsx and MapChrome.tsx. Legend rows and
keyboard-shortcut rows move to data/architecture-map.ts, already the declared
single source of truth for this page.

ArchitectureMap.tsx is now 116 lines. Behavior-preserving: every id, class and
data attribute the interaction layer queries survives byte-identical, as does
the viewBox save and restore handoff from the no-JS baseline.
